+++ src/main/java/com/jointpurchases/domain/point/model/dto/GetPoint.java
@@ -7,5 +7,5 @@ import lombok.Getter;
 @Getter
 @Builder
 public class GetPoint {
-    private Long currentPoint;
+    private Integer currentPoint;
 }
